+++ app/src/main/java/nextstep/jwp/controller/DefaultController.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.controller;
 
-import nextstep.jwp.framework.http.HttpRequest;
-import nextstep.jwp.framework.http.HttpResponse;
-import nextstep.jwp.framework.http.HttpStatus;
+import nextstep.jwp.framework.http.common.HttpStatus;
+import nextstep.jwp.framework.http.request.HttpRequest;
+import nextstep.jwp.framework.http.response.HttpResponse;
 
 public class DefaultController extends AbstractController {
