+++ app/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -4,9 +4,9 @@ import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.DBNotFoundException;
 import nextstep.jwp.exception.PasswordMismatchException;
-import nextstep.jwp.framework.http.HttpRequest;
-import nextstep.jwp.framework.http.HttpResponse;
-import nextstep.jwp.framework.http.HttpStatus;
+import nextstep.jwp.framework.http.common.HttpStatus;
+import nextstep.jwp.framework.http.request.HttpRequest;
+import nextstep.jwp.framework.http.response.HttpResponse;
 import nextstep.jwp.framework.http.session.HttpSession;
 import nextstep.jwp.framework.http.session.HttpSessions;
 import nextstep.jwp.model.User;
