+++ app/src/main/java/nextstep/jwp/framework/http/common/HttpBody.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.framework.http;
+package nextstep.jwp.framework.http.common;
 
 import java.util.Map;
