+++ app/src/test/java/nextstep/jwp/framework/http/QueryParamsTest.java
@@ -1,9 +1,9 @@
 package nextstep.jwp.framework.http;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.Map;
+import nextstep.jwp.framework.http.common.QueryParams;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
